+++ platforms/nuttx/src/px4/common/CMakeLists.txt
@@ -38,6 +38,7 @@ if(NOT PX4_BOARD MATCHES "io-v2")
 		board_crashdump.c
 		board_dma_alloc.c
 		board_fat_dma_alloc.c
+		cdc_acm_check.cpp
 		console_buffer.cpp
 		cpuload.cpp
 		gpio.c
